+++ demos/src/basic.rs
@@ -25,12 +25,12 @@ fn main() {
     .unwrap();
     let query_u8 = encoded.encode_query(&query);
 
-    let indexes = (0..vectors_count).collect::<Vec<_>>();
+    let indexes = (0..vectors_count as u32).collect::<Vec<_>>();
     let mut scores = vec![0.0; vectors_count];
     encoded.score_points(&query_u8, &indexes, &mut scores);
 
     for i in 0..vectors_count {
-        let score = encoded.score_point(&query_u8, i);
+        let score = encoded.score_point(&query_u8, i as u32);
         let score2 = scores[i];
         let orginal_score = dot_similarity(&query, &vector_data[i]);
         assert!((score - orginal_score).abs() < error);
@@ -38,7 +38,7 @@ fn main() {
     }
 
     for i in 0..vectors_count {
-        let score = encoded.score_internal(0, i);
+        let score = encoded.score_internal(0, i as u32);
         let orginal_score = dot_similarity(&vector_data[0], &vector_data[i]);
         assert!((score - orginal_score).abs() < error);
     }
